vue blog

+++ src/components/board/article/content.vue
@@ -12,34 +12,31 @@
       <v-toolbar color="transparent" dense flat>
         <v-toolbar-title>
           <v-btn
-            color="info"
+            color="primary"
             depressed
             small
             class="mr-4"
+            outlined
             @click="goCategory"
           >
             {{article.category}}
             <v-icon v-if="!category" right>mdi-menu-right</v-icon>
           </v-btn>
-          <template v-if="!$vuetify.breakpoint.xs">
-            <v-icon color="error" left v-if="newCheck(article.updatedAt)">mdi-fire</v-icon>
-            <span v-text="article.title"></span>
-          </template>
         </v-toolbar-title>
         <v-spacer/>
         <template v-if="(fireUser && fireUser.uid === article.uid) || (user && user.level === 0)">
-          <v-btn @click="articleWrite" icon><v-icon>mdi-pencil</v-icon></v-btn>
-          <v-btn @click="remove" icon><v-icon>mdi-delete</v-icon></v-btn>
+          <v-spacer/>
+          <v-btn @click="articleWrite" icon color="primary"><v-icon>mdi-pencil</v-icon></v-btn>
+          <v-btn @click="remove" icon color="error"><v-icon>mdi-delete</v-icon></v-btn>
         </template>
         <v-btn @click="back" icon><v-icon>mdi-close</v-icon></v-btn>
       </v-toolbar>
       <v-divider/>
-      <v-card-title v-if="$vuetify.breakpoint.xs">
-        <v-icon color="error" left v-if="newCheck(article.updatedAt)">mdi-fire</v-icon>
-        <span v-text="article.title"></span>
-      </v-card-title>
+      <v-card-subtitle class="text--primary body-1">
+        <display-title :item="article"/>
+      </v-card-subtitle>
       <v-card-text>
-        <viewer v-if="content" :initialValue="content"></viewer>
+        <viewer v-if="content" :initialValue="content" @load="onViewerLoad" :options="tuiOptions"></viewer>
         <v-container v-else>
           <v-row justify="center" align="center">
             <v-progress-circular indeterminate></v-progress-circular>
@@ -66,26 +63,38 @@
         <display-user :user="article.user"></display-user>
       </v-card-actions>
       <v-card-actions>
-        <!-- <v-chip small label outlined color="info" class="mr-2" v-for="tag in article.tags" :key="tag" v-text="tag"></v-chip> -->
         <v-spacer/>
-        <v-sheet class="mr-4">
-          <v-icon left :color="article.readCount ? 'info' : ''">mdi-eye</v-icon>
-          <span class="body-2">{{article.readCount}}</span>
-        </v-sheet>
-        <v-sheet class="mr-0">
-          <v-icon left :color="article.commentCount ? 'info' : ''">mdi-comment</v-icon>
-          <span class="body-2">{{article.commentCount}}</span>
-        </v-sheet>
-        <v-btn text @click="like">
-          <v-icon left :color="liked ? 'success' : ''">mdi-thumb-up</v-icon>
-          <span class="body-2">{{article.likeCount}}</span>
+        <display-count :item="article" :column="false"></display-count>
+      </v-card-actions>
+      <v-card-actions>
+        <v-spacer/>
+        <v-btn rounded @click="like" :color="liked ? 'success' : ''">
+          <v-icon left>mdi-thumb-up-outline</v-icon>
+          좋아요
+          <!-- <span class="body-2">{{article.likeCount}}</span> -->
         </v-btn>
       </v-card-actions>
       <v-card-text>
-        <v-row justify="end">
+        <v-row justify="start" align="center" class="px-4">
+          <v-btn
+            color="primary"
+            depressed
+            small
+            outlined
+            class="mr-4 mb-2"
+            @click="goCategory"
+          >
+            {{article.category}}
+            <v-icon right>mdi-menu-right</v-icon>
+          </v-btn>
           <v-chip small label outlined color="info" class="mr-2 mb-2" v-for="tag in article.tags" :key="tag" v-text="tag"></v-chip>
         </v-row>
       </v-card-text>
+      <v-card-actions v-if="(fireUser && fireUser.uid === article.uid) || (user && user.level === 0)">
+        <v-spacer/>
+        <v-btn @click="articleWrite" text color="primary"><v-icon left>mdi-pencil</v-icon>수정</v-btn>
+        <v-btn @click="remove" text color="error"><v-icon left>mdi-delete</v-icon>삭제</v-btn>
+      </v-card-actions>
       <v-divider/>
       <v-card-actions class="py-0">
         <v-row no-gutters>
@@ -106,26 +115,31 @@
       <display-comment :article="article" :docRef="ref"></display-comment>
     </v-card>
   </v-container>
-
 </template>
 <script>
 import axios from 'axios'
 import DisplayTime from '@/components/display-time'
 import DisplayComment from '@/components/display-comment'
 import DisplayUser from '@/components/display-user'
-import newCheck from '@/util/newCheck'
+import DisplayTitle from '@/components/display-title'
+import DisplayCount from '@/components/display-count'
+import addYoutubeIframe from '@/util/addYoutubeIframe'
 
 export default {
-  components: { DisplayTime, DisplayComment, DisplayUser },
-  props: ['boardId', 'articleId', 'category', 'tag'],
+  components: { DisplayTime, DisplayComment, DisplayUser, DisplayTitle, DisplayCount },
+  props: ['boardId', 'articleId', 'action', 'category', 'tag'],
   data () {
     return {
+      tuiOptions: {
+        linkAttribute: {
+          target: '_blank'
+        }
+      },
       content: '',
       ref: null,
       unsubscribe: null,
       article: null,
       doc: null,
-      newCheck,
       loaded: false
     }
   },
@@ -146,6 +160,9 @@ export default {
       this.subscribe()
     },
     articleId () {
+      this.subscribe()
+    },
+    action () {
       this.subscribe()
     }
   },
@@ -187,6 +204,14 @@ export default {
       this.$router.push({ path: this.$route.path, query: { action: 'write' } })
     },
     async remove () {
+      const r = await this.$swal.fire({
+        title: '정말 삭제하시겠습니까?',
+        text: '삭제 후 되돌릴 수 없습니다.',
+        icon: 'error',
+        // confirmButtonText: 'Cool',
+        showCancelButton: true
+      })
+      if (!r.value) return
       await this.ref.delete()
     },
     back () {
@@ -227,7 +252,7 @@ export default {
       let sn
       if (arrow < 0) sn = await ref.endBefore(this.doc).limitToLast(1).get()
       else sn = await ref.startAfter(this.doc).limit(1).get()
-      if (sn.empty) return this.$toast.info('더이상 페이지가 없습니다')
+      if (sn.empty) throw Error('더이상 페이지가 없습니다')
       const doc = sn.docs[0]
       const us = this.$route.path.split('/')
       us.pop()
@@ -239,6 +264,9 @@ export default {
       const us = this.$route.path.split('/')
       us.pop()
       this.$router.push({ path: us.join('/'), query: { category: this.article.category } })
+    },
+    onViewerLoad (v) {
+      addYoutubeIframe(v.preview.el, this.$vuetify.breakpoint)
     }
   }
 }
